chore: split into lib/binary crate
